declare convert yaml in hpp
